requirements.txt
streamlit
google-generativeai
Pillow
streamlit-javascript
